requirements.txt
click>=8.1
pdfminer.six>=202312
pytesseract>=0.3.13
pdf2image>=1.17
Pillow>=10.0
rapidfuzz>=3.9
regex>=2024.5.15
python-dateutil>=2.9
PyYAML>=6.0.1
pytest>=8.0
